ELG-182 Address review comments etc
Scale gnss values within the protocol layer
Use AES_BLOCKLEN define for clarity where rounding was done
Hold vpe/hpe as int everywhere
rename add_cache/get_cache
Make March 1st 2019 timestamp a define

+++ libel/beacons.h
@@ -135,9 +135,9 @@ typedef union beacon {
 typedef struct gps {
     double lat;
     double lon;
-    float hpe;
+    uint32_t hpe;
     float alt; // altitude
-    float vpe;
+    uint32_t vpe;
     float speed;
     float bearing;
     uint32_t nsat;
